Broader support for KeyChain.EXTRA_NAME
Previously EXTRA_NAME only supplied the default name when used with
KeyChain.EXTRA_CERTIFICATE. It now also provides the default name with
KeyChain.EXTRA_PKCS12, overriding the value found within the PKCS#12
file. In addition, will be used as the default alias name when
credentials are installed from external storage as opposed to via
an extra.

Bug: 6129421
Change-Id: I63c06a866fc62bfa56636011b766d3dcae7764bd