Add spinner to select Wi-Fi
For the regular installation path, we now need to show a spinner so we
don't regress from previous behavior where we can install certs for wifi
from the main installation screen.

You can also get to this from the Wi-Fi AP list, so only show it when
the intended use is ambiguous.

Bug: 8600545
Change-Id: I8e9c905ff84142ed7c1f50bf77eb3fb5574f8df1